afsocket: move socket plumbing to lib/socket
