+++ lib/CMakeLists.txt
@@ -38,6 +38,7 @@ add_subdirectory(secret-storage)
 add_subdirectory(logthrsource)
 add_subdirectory(logthrdest)
 add_subdirectory(signal-slot-connector)
+add_subdirectory(socket)
 
 set(LIB_SUBDIR_HEADERS
     ${ACK_TRACKER_HEADERS}
@@ -66,6 +67,7 @@ set(LIB_SUBDIR_HEADERS
     ${LOGTHRDEST_HEADERS}
     ${LOGTHRSOURCE_HEADERS}
     ${SIGNAL_SLOT_CONNECTOR_HEADERS}
+    ${SOCKET_HEADERS}
     ${ADT_HEADERS}
     ${CMAKE_CURRENT_BINARY_DIR}/filter/filter-expr-grammar.h
     ${CMAKE_CURRENT_BINARY_DIR}/filterx/filterx-grammar.h
@@ -288,6 +290,7 @@ set(LIB_SOURCES
     ${LOGTHRDEST_SOURCES}
     ${LOGTHRSOURCE_SOURCES}
     ${SIGNAL_SLOT_CONNECTOR_SOURCES}
+    ${SOCKET_SOURCES}
     ${ADT_SOURCES}
     ${PROJECT_BINARY_DIR}/lib/cfg-grammar.c
     ${PROJECT_BINARY_DIR}/lib/block-ref-grammar.c
@@ -469,6 +472,7 @@ install(FILES ${CSV_SCANNER_HEADERS} DESTINATION include/syslog-ng/scanner/csv-s
 install(FILES ${LOGTHRDEST_HEADERS} DESTINATION include/syslog-ng/logthrdest)
 install(FILES ${LOGTHRSOURCE_HEADERS} DESTINATION include/syslog-ng/logthrsource)
 install(FILES ${SIGNAL_SLOT_CONNECTOR_HEADERS} DESTINATION include/syslog-ng/signal-slot-connector)
+install(FILES ${SOCKET_HEADERS} DESTINATION include/syslog-ng/socket)
 
 set(TOOLS
     merge-grammar.py
