+++ lib/socket/transport-mapper.c
