+++ modules/afsocket/CMakeLists.txt
@@ -6,22 +6,12 @@ set(AFSOCKET_SOURCES
     afsocket-dest.c
     afsocket-dest.h
     afsocket-signals.h
-    socket-options.c
-    socket-options.h
-    transport-mapper.c
-    transport-mapper.h
-    afinet.c
-    afinet.h
     afinet-source.c
     afinet-source.h
     afinet-dest-failover.c
     afinet-dest-failover.h
     afinet-dest.c
     afinet-dest.h
-    socket-options-inet.c
-    socket-options-inet.h
-    transport-mapper-inet.c
-    transport-mapper-inet.h
     afunix-source.c
     afunix-source.h
     afunix-dest.c
