+++ modules/afsocket/Makefile.am
@@ -6,12 +6,6 @@ modules_afsocket_libafsocket_la_SOURCES	=	\
 	modules/afsocket/afsocket-dest.c		\
 	modules/afsocket/afsocket-dest.h		\
 	modules/afsocket/afsocket-signals.h		\
-	modules/afsocket/socket-options.c        	\
-	modules/afsocket/socket-options.h		\
-	modules/afsocket/transport-mapper.c		\
-	modules/afsocket/transport-mapper.h		\
-	modules/afsocket/afinet.c			\
-	modules/afsocket/afinet.h			\
 	modules/afsocket/afinet-source.c		\
 	modules/afsocket/afinet-source.h		\
 	modules/afsocket/afinet-dest.c			\
@@ -20,12 +14,8 @@ modules_afsocket_libafsocket_la_SOURCES	=	\
 	modules/afsocket/logproto-spoof-source-client.h	\
 	modules/afsocket/afinet-dest-failover.c		\
 	modules/afsocket/afinet-dest-failover.h		\
-	modules/afsocket/socket-options-inet.c		\
-	modules/afsocket/socket-options-inet.h   	\
 	modules/afsocket/socket-options-unix.c		\
 	modules/afsocket/socket-options-unix.h   	\
-	modules/afsocket/transport-mapper-inet.c 	\
-	modules/afsocket/transport-mapper-inet.h 	\
 	modules/afsocket/afunix-source.c		\
 	modules/afsocket/afunix-source.h		\
 	modules/afsocket/afunix-dest.c			\
